# openapi.json
{
  "components": {
    "parameters": {
...
      "Limit": {
        "description": "Maximum number of items. The default is 15.",
        "example": 20,
        "in": "query",
        "name": "limit",
        "required": false,
        "schema": {
          "default": 15,
          "minimum": 1,
          "type": "integer"
        }
      },
      "MembershipFilter": {
        "description": "Filters by membership identifier.",
        "example": "monthly",
        "in": "query",
        "name": "membership",
        "required": false,
        "schema": {
          "maxLength": 40,
          "minLength": 3,
          "type": "string"
        }
      },
      "MembershipId": {
        "description": "Unique membership identifier.",
        "example": "monthly",
        "in": "path",
        "name": "membership_id",
        "required": true,
        "schema": {
          "maxLength": 40,
          "minLength": 3,
          "type": "string"
        }
      },
      "Month": {
        "description": "Statistics month (1 to 12).",
        "example": 8,
        "in": "query",
        "name": "month",
        "required": false,
        "schema": {
          "maximum": 12,
          "minimum": 1,
          "type": "integer"
        }
      },
...
      "NextCursor": {
        "description": "Cursor returned by the previous page.",
        "example": "monthly",
        "in": "query",
        "name": "next_cursor",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "UserId": {
        "description": "Numeric user identifier.",
        "example": 42,
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "minimum": 1,
          "type": "integer"
        }
      },
      "Year": {
        "description": "Statistics year.",
        "example": 2026,
        "in": "query",
        "name": "year",
        "required": false,
        "schema": {
          "minimum": 1,
          "type": "integer"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "content": {
          "application/json": {
            "example": {
              "message": "Validation error"
            },
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Invalid request or validation error."
      },
      "Forbidden": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "The manager lacks permission or the account is disabled."
      },
      "NotFound": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Resource not found."
      },
      "Unauthorized": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description": "Missing, invalid, or expired token."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "errors": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "Membership": {
        "properties": {
          "days": {
            "minimum": 1,
            "type": "integer"
          },
          "description": {
            "default": "None",
            "type": "string"
          },
          "id": {
            "maxLength": 40,
            "minLength": 3,
            "type": "string"
          },
          "price": {
            "minimum": 1,
            "type": "number"
          }
        },
        "required": [
          "id",
          "days",
          "price"
        ],
        "type": "object"
      },
      "MembershipPage": {
        "properties": {
          "has_next": {
            "type": "boolean"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/Membership"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "items",
          "next_cursor",
          "has_next"
        ],
        "type": "object"
      },
      "MessageResponse": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "StatisticInfo": {
        "properties": {
          "amount": {
            "type": "number"
          },
          "id": {
            "type": "string"
          },
          "membership": {
            "type": "string"
          },
          "month": {
            "maximum": 12,
            "minimum": 1,
            "type": "integer"
          },
          "sold_memberships": {
            "minimum": 0,
            "type": "integer"
          },
          "year": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "year",
          "month",
          "membership",
          "sold_memberships",
          "amount"
        ],
        "type": "object"
      },
...
      "UpdateMembership": {
        "minProperties": 1,
        "properties": {
          "days": {
            "minimum": 1,
            "type": [
              "integer",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "minimum": 1,
            "type": [
              "number",
              "null"
            ]
          }
        },
        "type": "object"
      },
      "UpdateMembershipUser": {
        "properties": {
          "membership": {
            "maxLength": 20,
            "minLength": 3,
            "type": "string"
          }
        },
        "required": [
          "membership"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AdminToken": {
        "description": "JWT issued by `POST /admin/sign-in`. Send it directly, without `Bearer`. Roles are checked on the server: MIN/MID/MAX for general operations and MAX for privileged administration.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {
      "name": "GymSystem"
    },
    "description": "Gym management API. It provides administration of managers, memberships, users, and sales statistics. Protected operations expect the JWT directly in the `Authorization` header, without the `Bearer` prefix.",
    "title": "GymSystem API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/membership": {
      "get": {
        "description": "Lists plans ordered by identifier. Available to MIN, MID, and MAX managers.",
        "operationId": "getMemberships",
        "parameters": [
          {
            "$ref": "#/components/parameters/NextCursor"
          },
          {
            "$ref": "#/components/parameters/Limit"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipPage"
                }
              }
            },
            "description": "Membership page."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "List memberships",
        "tags": [
          "Membership"
        ]
      },
      "post": {
        "description": "Creates a new plan. The identifier must be unique. Requires MAX.",
        "operationId": "createMembership",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "days": 30,
                "description": "Monthly access",
                "id": "monthly",
                "price": 499.99
              },
              "schema": {
                "$ref": "#/components/schemas/Membership"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Membership created."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Create a membership",
        "tags": [
          "Membership"
        ]
      }
    },
    "/membership/info/{membership_id}": {
      "delete": {
        "description": "Deletes the specified plan. Requires MAX.",
        "operationId": "deleteMembership",
        "parameters": [
          {
            "$ref": "#/components/parameters/MembershipId"
          }
        ],
        "responses": {
          "204": {
            "description": "Membership deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Delete a membership",
        "tags": [
          "Membership"
        ]
      },
      "get": {
        "description": "Returns the requested plan details. Available to MIN, MID, and MAX managers.",
        "operationId": "getMembershipInfo",
        "parameters": [
          {
            "$ref": "#/components/parameters/MembershipId"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            },
            "description": "Membership details."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Get a membership",
        "tags": [
          "Membership"
        ]
      },
      "patch": {
        "description": "Updates one or more plan fields. Requires MAX.",
        "operationId": "updateMembershipInfo",
        "parameters": [
          {
            "$ref": "#/components/parameters/MembershipId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembership"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Membership updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Update a membership",
        "tags": [
          "Membership"
        ]
      }
    },
    "/statistics": {
      "get": {
        "description": "Queries sales grouped by year, month, and membership. All filters are optional. Requires MAX.",
        "operationId": "getStatistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/Year"
          },
          {
            "$ref": "#/components/parameters/Month"
          },
          {
            "$ref": "#/components/parameters/MembershipFilter"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StatisticInfo"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Statistics found."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Get statistics",
        "tags": [
          "Statistics"
        ]
      }
    },
    "/statistics/current": {
      "get": {
        "description": "Returns statistics for the current month or for a specific membership. Requires MAX.",
        "operationId": "getCurrentMonthStatistics",
        "parameters": [
          {
            "$ref": "#/components/parameters/MembershipFilter"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/StatisticInfo"
                    },
                    {
                      "items": {
                        "$ref": "#/components/schemas/StatisticInfo"
                      },
                      "type": "array"
                    }
                  ]
                }
              }
            },
            "description": "A single statistic when filtering by membership; otherwise, an array of statistics."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Get current-month statistics",
        "tags": [
          "Statistics"
        ]
      }
    },
...
    "/user/membership/{id}": {
      "patch": {
        "description": "Changes the user's plan, recalculates the next payment date, and records the sale.",
        "operationId": "updateUserMembership",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserId"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipUser"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            },
            "description": "Membership updated."
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "AdminToken": []
          }
        ],
        "summary": "Change a user's membership",
        "tags": [
          "User"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local server",
      "url": "http://127.0.0.1:8080"
    }
  ],
...
}